+++ diff-coverage/src/funcTest/resources/test-excludes-classes-project/src/main/java/com/excludes/UncoveredClass.java
@@ -1,0 +1,6 @@
+package com.excludes;
+public class UncoveredClass {
+    public void method() {
+        System.out.println(1);
+    }
+}