+++ diff-coverage/src/funcTest/resources/test-excludes-classes-project/src/main/java/com/excludes/sub/SubUncoveredClass.java
@@ -1,0 +1,6 @@
+package com.excludes.sub;
+public class SubUncoveredClass {
+    public void method() {
+        System.out.println(1);
+    }
+}